Move some headers. Plugin: remove ecmcRefs.
