+++ devEcmcSup/main/ecmcMainThread.cpp
@@ -550,12 +550,9 @@ int setAppModeRun(int mode) {
   }
 
   // Plugins
-  // populate pluginDataRefs
-  pluginDataRefs.ecmcAsynPort = asynPort;
-  pluginDataRefs.sampleTimeMS = 1 / mcuFrequency;
   for(int i=0; i < ECMC_MAX_PLUGINS; ++i) {
     if(plugins[i]) {
-      errorCode = plugins[i]->exeEnterRTFunc(&pluginDataRefs);
+      errorCode = plugins[i]->exeEnterRTFunc();
       if(errorCode){
         return errorCode;
       }
