+++ devEcmcSup/motion/ecmcAxisBase.cpp
@@ -17,6 +17,7 @@
 #include <new>
 #include <iostream>
 
+#include "../main/ecmcErrorsList.h"
 
 /**
  *
